requirements.txt
pytest
playwright
requests
python-dotenv 
locust